requirement.txt
# machine learning
transformers==4.19.2
datasets==2.2.2
torchaudio==0.11.0
scikit-learn==1.1.1

# data preprocessing
pandas==1.4.2

# Additional package
evaluate
torchsummary
torchinfo